requirements.txt
fastapi
uvicorn
jinja2
httpx
python-dotenv
pydantic
aiofiles
gitpython
transformers
torch 
python-multipart